Removed the ellipse from signature
You pass the parameters to mean() function but you don't store them. So, when you call the function with different parameters, instead of recalculating the mean you pull it from the cache. If you call mean with or without NA removing, you'll get incorrect results using this cached mean method. Therefore, it's better to remove the ellipse. 

Alternatively you could implement handling of parameter changes but that' more code and is not related to the subject of this example.