Completed assignment.
This commit includes the cachematrix.R file, as specified by the
assignment. The file includes the makeCacheMatrix function, and the
cacheSolve function. Both functions work as intended to the best of my
knowledge.

There is also an example.r file, which is the sample code provided in
the assignment. I used this code to better understand the behavior we
were to replicate in the assignment.