R Programming Assignment -2, Lexical Scoping (Week-3)
There are two of  functions in "makeCacheMatrix"  and cacheSolve.
makeCacheMatrix creates a special "matrix" object that can cache its
inverse. cacheSolve which computes the inverse of the special "matrix"
returned by makeCacheMatrix above. If the inverse has already been
calculated (and the matrix has not changed), then the cachesolve
retrieve the inverse from the cache.