Create RUNZE WU
Caching the Inverse of a Matrix:
The first function is about Matrix inversion is usually a costly computation and there may be some benefits to caching the inverse of a matrix rather than compute it repeatedly. Below are a pair of functions that are used to create a special object that stores a matrix and caches its inverse. This function creates a special "matrix" object that can cache its inverse.
The second function computes the inverse of the special "matrix" created by makeCacheMatrix above. If the inverse has already been calculated and the matrix has not changed), then it should retrieve the inverse from the cache.