Update cachematrix.R
makeCacheMatrix is the function that creates a special "matrix" object that can cache its inverse. cacheSolve is the function which computes the inverse of the special "matrix" returned by makeCacheMatrix above. If the inverse has already been calculated (and the matrix has not changed), then the cacheSolve should retrieve the inverse from the cache.